apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.29.0",
  "httpx>=0.27.0",
  "pyjwt[crypto]>=2.8.0",
  "sqlalchemy[asyncio]>=2.0.0",
  "asyncpg>=0.29.0",
  "alembic>=1.13.0",
  "arq>=0.25.0",
  "redis>=5.0.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.0",
  "cryptography>=42.0.0",
  "tiktoken>=0.7.0",
  "anthropic>=0.28.0",
  "unidiff>=0.7.5",
  "pyyaml>=6.0.3",
  "tree-sitter-language-pack>=1.6.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.0",
]

[tool.pytest.ini_options]
markers = [
  "live_github: integration test that hits the real GitHub API",
]
